Migration

+++ actions/MassDeleteAction.php
@@ -3,6 +3,7 @@ namespace dezmont765\yii2bundle\actions;
 
 use dezmont765\yii2bundle\components\Alert;
 use Exception;
+use yii\web\ForbiddenHttpException;
 
 /**
  * Created by PhpStorm.
@@ -25,6 +26,9 @@ class MassDeleteAction extends MainAction
                 try {
                     $model_class = $this->getModelClass();
                     $model = $this->controller->findModel($model_class, $key);
+                    if(!$this->controller->checkAccess($this->permission, ['model' => $model])) {
+                        throw new ForbiddenHttpException('You can\'t delete this item');
+                    }
                     if($model) {
                         if($model->delete()) {
                             Alert::addSuccess($this->success_message);
